Relocated damage from weapons to gameplay entities, added a condition for each temporary buff to avoid potential errors

+++ Assets/Scripts/SO/GameplayEntityStats/GamePlayEntityStats.cs
@@ -4,9 +4,9 @@ public class GamePlayEntityStats : ScriptableObject
 {
     [SerializeField] private int _healthValue;
     [SerializeField] private float _speedValue;
-    [SerializeField] private float _damageValue;
+    [SerializeField] private int _damageValue;
 
     public int HealthValue => _healthValue;
     public float SpeedValue => _speedValue;
-    public float DamageValue => _damageValue;
+    public int DamageValue => _damageValue;
 }
